refactor: left button = quit
