+++ clientnew/src/states.rs
@@ -4,13 +4,25 @@ use bevy::prelude::*;
 pub enum GameState {
     #[default]
     MainMenu,
+    Connecting,
+    CreateServer,
     InGame,
+    Quit,
 }
 
 pub struct StatePlugin;
 
 impl Plugin for StatePlugin {
     fn build(&self, app: &mut App) {
-        app.add_state::<GameState>();
+        app
+            .add_state::<GameState>()
+            .add_systems(OnEnter(GameState::Quit), quit_event)
+            ;
     }
+}
+
+fn quit_event(
+     mut app_exit_events: ResMut<Events<bevy::app::AppExit>>
+) {
+    app_exit_events.send(bevy::app::AppExit);
 }
